fixed a bug where the tiles would appear to have all 4 walls even though they were right next to tiles owned by the player. Also got the project ready for another set of tests. Next is to add feedback to the player when things happen and when they act.

+++ ZombieBoardGame/Assets/Scripts/Menus/ActionPanels/SubPanels/RaidSubPanel.cs
@@ -20,7 +20,7 @@ public class RaidSubPanel : ActionSubPanel
         }
         float temp = ((numPeople * PeopleCapabilities.numZombiesOnePersonCanHandle)
             + (numberOfArmedPeople * (PeopleCapabilities.weaponKillModifier))) / Map.instance.getTileAt(currentXLoc, currentYLoc).numberOfZombiesOccupying;
-        ChanceText.text = "" + Mathf.Clamp(temp, 0, 1f) * 100 + "%";
+        ChanceText.text = "" + (int)(Mathf.Clamp(temp, 0, 1f) * 100) + "%";
     }
 
     public override void submitInfoToActionList()
